Fixed LogLevel logic bug.
The logger was logging any message equal to or below the console or log file LogLevel, when it should've been logging only messages equal to or above. That way you can have a log with LogLevel = LOGERROR, which only contains messages of LogLevel LOGERROR, or LOGFATALERROR. Easier to find bugs and issues in logs. Less manual sorting.