Added basic script detection and running
The script manager now scans the main scripts folder (no subfolder scanning yet) and adds each unique lua file to a list of scripts which can be run. Files without the .lua extension are skipped. The main overlay also now has a scripts section which shows a list of these scripts and allows the user to run them.