Added a few more rfg functions to test rl_draw struct
My tests of this failed, with rl_draw::string() causing the game to freeze. For now that's all I am doing, but later on it'd be a good thing to mess with instead of using an external lib. From setting breakpoints I know that rl_draw::tristrip_2d_begin() is called in the pause menu. Not sure about the other functions.