Started reorganization of overlay code
The overlay code now has the GuiSystem class which manages the state and behavior of all other gui elements/windows/menus. The idea here is to have each window be it's own class instead of just having one massive file like before with MainOverlay. There's still a lot of work to do here. Need to cut down MainOverlay more and make separate classes for or remove each of the elements it contains. I'd also like to switch MainOverlay into a help window. This would pop up the first time the player opens the overlay, and show them the basics.