Set warning level to /W4 . Disabled warnings for external code
Any code included using brackets < > is considered to be external by the compiler now. So, to avoid dozens of warnings from new libraries make sure to include them with brackets. not quotation marks. Having higher warning levels should catch simple code mistakes such as semicolons at the end of an if statement. Note that a pragma had to be use to silence a warning in the kiero library.