Created top menu bar overlay gui.
Will be above console and allow access to all of the menus and important game info. Overlay will start off with only this menu bar and the about window and people can open any window they want from there. Will likely be useful to update imgui to the docking branch to allow better sorting and space usage for all menus.