Added open/close logic to all gui menus
I somehow forgot to add logic for checking if menus should be open or closed on all gui elements. This fixes that, they should all close properly. Now, the main menu bar just needs to be populated with submenus and buttons for each so they can be opened again if closed.