Added explosion spawning and ambient light vars to tweaks gui
Can spawn explosion with middle mouse button. Change spawn settings in the tweaks gui. There are issues with making normal explosion effects appear, so I'll try to come up with a way to iterate through the explosions list and copy those values instead of needing to always manually set it up. Ambient light changes work but there seems to be some other skylight or ambient light that also effects lighting. Probably an rl_light, might be a func with tod in it that lets you change the color or tod