Initial pass on core lua library
Includes basic enums and types. This is mimicking the design of MWSE2 and it's core library. The goal is to put as much as possible on the lua side to make things easier to edit. Will need to add a way to reload the core lib when the script loader is injected for testing purposes.