Temporary scripting fixes for early release
Just going to perform a temporary fix with the scripting/sol build error so I can finish up this release for today. Going to split the release in two. Release what I have now, with some tweaks tonight, and then focus solely on scripting and minor bugfixes for the next release, which I plan on releasing much sooner.