Created BaseGui class.
All imgui gui's will be derived from this. It'll simplify adding new guis, so all that needs to be done is for a new derived class to be made and added to the GuiSystem vector. Removes the possibility of forgetting to add a given gui to one of several GuiSystem functions needed to make them work properly.