Improved gui improve structure
Before was basically a big chain of inclusions which could be a pain to manage at times and meant that if one gui class changed all of the ones above it in the include chain needed to be changed. Should speed up compilation in many cases and make it speedier to add new guis.