Moved input code into WndProcHook
Now the only job of the main thread after init and hooking is to check if it should close. The text editor still uses GetAsyncKeyState() since I couldn't get multi-key presses, like Ctrl+Shift+S for example to work with the normal input detection. Might need to use bools like "CtrlKeyDown" and split it into multiple checks so it can work across calls to the wndproc on the input queue.