GuiSystem object now a member of ProgramManager :wrench:
Still available as a pointer through Globals::Gui. This is much more preferred as now it's clear that ProgramManager owns the only instance of GuiSystem rather than having it in a odd spot hidden away in Hooks.h. Also nice since it's another spare pointer laying around that can be forgotten. There were previously tons of pointers to the same thing being updated which could've lead to issues if one was missed.