Add Utilities::General::ThreadedBeep
Partial solution to windows func Beep stopping execution on it's calling thread until it's sound is done. Need a better one as this still has a slight pause and has the problem of spawning a new thread each time it's called. When it's called too often it just completely breaks the beep functionality until the game crashes. So that's not good