Temporary fix for script editor crash
Could be due to me changing the colorizer to a c++ colorizer, (since it's more performant due to not using std::regex). Don't normally like to edit the text editor since it makes upgrading harder but this shouldn't be a huge thing to edit back in if the problem remains after updating it next time.