Moved around init code
The main goal with this was to make it so hooks are set while the threads are still suspended, hopefully letting us hook and patch early game code more easily. Doesn't seem to be working fully since replacing interface_init_load.cpeg_pc with the peg load wrapper hook still isn't working. Will have to think of better ways of doing this