Add hook to D3D11ResizeBuffers for debugging
Adding this as part of trying to debug the resize crash. Haven't managed to solve it yet even with the info from this hook. Seems that the D3D11 hooking code in KeenGraphicsBeginFrameHook is mucking up the D3D11 state of the game and causing the crash, since the inputs to the ResizeBuffer call are identical whether or not the overlay hooks are active. Keeping this hook in because it will be useful for future debugging and graphical changes.