Add Appveyor build scripts
The actual build scripts are set through the Appveyor web gui currently since it was easier than setting up the yaml file. Adding them here so they're publically available. In the future it might be a good idea to move all config to the repo if possible. According to this link we might need to have the yml on a separate github gist though: https://www.appveyor.com/docs/build-configuration/#generic-git-repositories-and-yaml