Attempt to remedy bug with finding explosion and effect presets
This code basically checks if the memory values (like Globals::NumExplosionInfos) that it sets are correct before continuing. If they are incorrect it waits 500ms and tries again. It will do this 30 times (15 seconds) before it gives up and shuts down the script loader and game. This also adds more logging during that process that should hopefully help find any more bugs if the issue persists.