Fixed a crash during vehicle spawning that sometimes occurs
Unsure if this is an issue with the base game. I've only been able to consistently reproduce it with the script loader active, and a debugger attached. The problem occurs when a nullptr is passed to object_spawn_vehicle and used without being checked. The fix is to simply hook the function and check if the value is null before the function runs. If it's null, the hook returns the spawn failure enum SPAWN_STATUS_FAILURE.