Change first person cam to move to third person by default with vehicles
Can be toggled in the cam settings menu, but going to third person is the default option. This is here because the first person cam doesn't work well with vehicles as is. The vehicles appear to shake violently from a first person view as they move along, which is less than ideal