Add alternatives to common objects like rfg.ActivePlayer
Can use `Player` instead of `rfg.ActivePlayer`
Can use `World` instead of `rfg.ActiveWorld`
Can use `PhysicsWorld` instead of `rfg.ActivePhysicsWorld`
Can use `World.Objects` instead of `rfg.ActiveWorld.AllObjects`
The old way is still available, but this is much better since it's quicker to type and avoids some confusion. In a few months or so these will likely be deprecated. It'd be nice if there was a way to provide warnings to people in the script editor like "this will be deprecated in a month" when they use certain objects.