Reorganize default/included scripts
Moved all the scripts that come with the RSL to their own subfolder, deleted a few, and updated the others with more comments, and a few syntax changes and bugfixes. The main goal was to de-clutter the scripts folder a bit so it's not immediately filled with tons of scripts.