Add Timer class and make use of it
This is a simple wrapper class around std::chrono::steady_clock which lets you avoid all the verbose syntax of that, and use simple "Timer.Start" and "Timer.Reset" funcs to track time that has passed. 

I also made use of the timer with the explosion spawning timer and the script editor run button. The run button previously slept the thread for 150ms, causing a distinct pause whenever a script was run. That's now gone since the non-blocking timer is used instead.

... Need to bind this to lua...