Initial work on editing accessing rfg menu information
Does not change the behavior of the RSL at all or add any new features yet. Just some internal work that will make editing rfg menus easier later on. There seems to be an issue with RSLs version of rfg::farrray<> since the test button that's commented out in the general tweaks gui code reads the arrays length as zero even though it is not zero.