Fix a crash in the script editor (Thanks Arrows)
Caused by editing a script file then trying to open another one and clicking any of the buttons on the context menu that popped up. Not sure of the extend of the issue and exactly what was happening but it seems to have been something caused by using std::partition to sort each file browser tree node to have folders at the start of each node children list. Replacing it with a hand written sorting function fixed this.