Add "Initialized" events
Called once on the first time a save is loaded. When this is called you can be sure that all values scripts might use are initialized and ready to use. Calling functions or using values before this event is triggered is unsafe, as you can't be sure that all values are ready for use. 

This lets you safely create scripts that automatically run by combining autorun scripts (ones named main.lua), and initialized events.