Fix bug where the WeaponInfo list would not be properly initialized
The issue only occurred in release builds. Debug builds were slow enough that the game had time to initialize the weapons info list before the RSL tried accessing it. In optimized release builds the RSL was too quick and didn't wait long enough for the list to be initialized. The fix makes the RSL wait for it to be initialized. This works for now, but doesn't look great. Need a better way to handle waiting for the game to initialize. Hooking it's init functions and waiting for them to finish should work.