Add "Mouse" and "Load" events
Mouse events are triggered every time the mouse is moved or the mouse wheel is scrolled. Will also add mouse click detection to this event in the future. Load events occur every single time a save is loaded (including when you die and respawn). This is useful for resetting things like user messages which get removed upon load/respawn. Will write a guide for events and document the existing ones in the next few days in between classes.