Update sol3 and temporarily disable classes causing compiler to run out of heap-space (Now compiles!)
This is an out of date test branch used to see if we can finally switch from sol2 to sol3. With VS2019 16.3.3 it now compiles with sol3, so we can switch! Though I'll probably wait since the latest version still has some usertypes that have to be converted and a few usertypes need to be changed so they don't need 60+GB of ram to compile