Add a bunch of solution explorer filters + Basic IPC code for teleporting player
The solution explorer filters makes navigating it much more manageable, and makes it easier to find files. 

IpcManager currently is only for teleporting the player to a position when received over a named pipe. No security measures are put in place for this since it's a position, not critical data.

This will later be used to have external tools like the zone editor teleport the player to a position to confirm what a zone object is instead of manually typing the position each time. Not sure on the other uses yet.