Add Hypodermic IoC container submodule and basic usage of it with IIpcManager
Boost is required by Hypodermic sadly. I used a precompiled version since compiling it was a pain for me. The download is large and I don't want to have to keep a copy around, so what I'll likely do is find the minimum required file/lib set for boost and upload it to another repo so it doesn't have to be stored in it's entirety or rebuilt on each commit.