Refactor RSL to make use of IoC Container
This made a lot of changes, so there's likely a few new bugs, and more refactoring to do.

Using IoC should make the code more modular and less spaghet. It also should make unit testing / mocking easier once that's added as that's something I want to add to the project in the near future.