Remove USE_LAUNCHER macro, add file overrides instead
Make an empty text file called RSL_No_Launcher.txt in the rfg folder to run the game with RSL enabled without showing the launcher. Make an empty text file called Vanilla_No_Launcher.txt to run the game without the launcher and without the RSL. These are overrides that skip the launcher for people who the launcher freezes for. It still does this sometimes and isn't reproducible, so these workarounds are here for now.