Add rfg.CreatePlayerSquad
Lets you spawn player squads. They follow the player around. Use the squad names from squads.xtbl. So for example,  the first squad in that file is `basic_squad_1`. To spawn that squad you'd do the following in lua:
`rfg.CreatePlayerSquad("basic_squad_1")`

There's currently no way to spawn squads independent of the player, but that should be coming soon if I can get it working.