[skip ci] Refactor Logger, reduce default log cache size, and optimize slightly
Refactors Logger to make the code much shorter and more concise. Switches LogData to be a std::deque<LogEntry> instead of a std::vector since each message is inserted at the front of LogData, which is very inefficient with a vector, but fine to do with a deque. Renamed LogData to LogCache. Reduced default max log cache size from 2500 to 1000 since it's more reasonable. Wanted to cut down on it's memory usage.