[skip ci] Script editor refactor
Cut ~100 LOC off the script editor and made the code much cleaner and easier to read. Made use of Path:: functions in a few places instead of manually splicing paths, which makes the code much less messy. I still think the popup code is a bit verbose and wonder if it could be shortened somehow.